Remove deprecated code from default code API.
Move perlin noise and array methods into PGL code.
