+++ praxiscore-code/src/main/java/org/praxislive/code/DefaultDelegateAPI.java
@@ -173,146 +173,6 @@ public interface DefaultDelegateAPI {
     }
 
     /**
-     * Extract a double from the Property's current Value, or zero if the value
-     * cannot be coerced.
-     *
-     * @param p
-     * @return
-     */
-    @Deprecated
-    public default double d(Property p) {
-        return p.getDouble();
-    }
-
-    /**
-     * Convert the provided Value into a double, or zero if the Value cannot be
-     * coerced.
-     *
-     * @param v
-     * @return
-     */
-    @Deprecated
-    public default double d(Value v) {
-        if (v instanceof PNumber) {
-            return ((PNumber) v).value();
-        } else {
-            return PNumber.from(v).orElse(PNumber.ZERO).value();
-        }
-    }
-
-    /**
-     * Parse the provided String into a double, or zero if invalid.
-     *
-     * @param s
-     * @return
-     */
-    @Deprecated
-    public default double d(String s) {
-        return d(PString.of(s));
-    }
-
-    /**
-     * Extract an int from the Property's current Value, or zero if the value
-     * cannot be coerced.
-     *
-     * @param p
-     * @return
-     */
-    @Deprecated
-    public default int i(Property p) {
-        return p.getInt();
-    }
-
-    /**
-     * Convert the provided Value into an int, or zero if the Value cannot be
-     * coerced.
-     *
-     * @param v
-     * @return
-     */
-    @Deprecated
-    public default int i(Value v) {
-        if (v instanceof PNumber) {
-            return ((PNumber) v).toIntValue();
-        } else {
-            return PNumber.from(v).orElse(PNumber.ZERO).toIntValue();
-        }
-    }
-
-    /**
-     * Parse the provided String into an int, or zero if invalid.
-     *
-     * @param s
-     * @return
-     */
-    @Deprecated
-    public default int i(String s) {
-        return i(PString.of(s));
-    }
-
-    /**
-     * Extract the Property's current value as a boolean. If the value cannot be
-     * coerced, returns false.
-     *
-     * @param p
-     * @return
-     */
-    @Deprecated
-    public default boolean b(Property p) {
-        return p.getBoolean();
-    }
-
-    /**
-     * Convert the provided Value into a boolean according to the parsing rules
-     * of {@link PBoolean}. If the Value cannot be coerced, returns false.
-     *
-     * @param v
-     * @return
-     */
-    @Deprecated
-    public default boolean b(Value v) {
-        if (v instanceof PBoolean) {
-            return ((PBoolean) v).value();
-        } else {
-            return PBoolean.from(v).orElse(PBoolean.FALSE).value();
-        }
-    }
-
-    /**
-     * Parse the given String into a boolean according to the parsing rules of
-     * {@link PBoolean}. If the String is invalid, returns false.
-     *
-     * @param s
-     * @return
-     */
-    @Deprecated
-    public default boolean b(String s) {
-        return b(PString.of(s));
-    }
-
-    /**
-     * Extract the Property's current value into a String representation.
-     *
-     * @param p
-     * @return
-     */
-    @Deprecated
-    public default String s(Property p) {
-        return p.get().toString();
-    }
-
-    /**
-     * Convert the provided Value into a String representation.
-     *
-     * @param v
-     * @return
-     */
-    @Deprecated
-    public default String s(Value v) {
-        return v.toString();
-    }
-
-    /**
      * Attempt to extract a {@link PArray} from the given Property. An empty
      * PArray will be returned if the property's value is not a PArray and
      * cannot be coerced.
